// src/api/models/Player.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Player
 */
export interface Player {
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    email?: string;
    /**
     * 
     * @type {number}
     * @memberof Player
     */
    teamId?: number;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    lastName?: string;
    /**
     * 
     * @type {Date}
     * @memberof Player
     */
    birthDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof Player
     */
    firstName?: string;
}

/**
 * Check if a given object implements the Player interface.
 */
export function instanceOfPlayer(value: object): boolean {
    return true;
}

export function PlayerFromJSON(json: any): Player {
    return PlayerFromJSONTyped(json, false);
}

export function PlayerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Player {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'email': json['email'] == null ? undefined : json['email'],
        'teamId': json['teamId'] == null ? undefined : json['teamId'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'birthDate': json['birthDate'] == null ? undefined : (new Date(json['birthDate'])),
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
    };
}

export function PlayerToJSON(value?: Player | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'email': value['email'],
        'teamId': value['teamId'],
        'lastName': value['lastName'],
        'birthDate': value['birthDate'] == null ? undefined : ((value['birthDate']).toISOString().substring(0,10)),
        'firstName': value['firstName'],
    };
}