// src/api/models/Team.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Team
 */
export interface Team {
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof Team
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    city?: string;
    /**
     * 
     * @type {Date}
     * @memberof Team
     */
    establishmentDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    imageUrl?: string;
}

/**
 * Check if a given object implements the Team interface.
 */
export function instanceOfTeam(value: object): boolean {
    return true;
}

export function TeamFromJSON(json: any): Team {
    return TeamFromJSONTyped(json, false);
}

export function TeamFromJSONTyped(json: any, ignoreDiscriminator: boolean): Team {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'id': json['id'] == null ? undefined : json['id'],
        'description': json['description'] == null ? undefined : json['description'],
        'city': json['city'] == null ? undefined : json['city'],
        'establishmentDate': json['establishmentDate'] == null ? undefined : (new Date(json['establishmentDate'])),
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
    };
}

export function TeamToJSON(value?: Team | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'id': value['id'],
        'description': value['description'],
        'city': value['city'],
        'establishmentDate': value['establishmentDate'] == null ? undefined : ((value['establishmentDate']).toISOString().substring(0,10)),
        'imageUrl': value['imageUrl'],
    };
}